Update the set of unique signatures when loading an existing log.
